feat: finish the longfor check in script
