--- Script/longfor/longfor-checkin-shadowrocket.sgmodule
@@ -1,25 +1,0 @@
-#!name=龙湖自动签到测试
-#!desc=龙湖APP自动签到功能模块，包含获取Token和定时任务
-#!system=ios
-# 使用方法：
-# 1. 将此配置导入 Shadowrocket
-# 2. 开启 MITM 和脚本功能
-# 3. 打开龙湖APP登录获取token
-# 4. 脚本会自动执行签到
-
-[MITM]
-hostname = gw2c-hw-open.longfor.com
-
-[Script]
-# 获取Token脚本
-龙湖签到Token = type=http-response,pattern=^https:\/\/gw2c-hw-open\.longfor\.com\/supera\/member\/api\/bff\/pages\/v1_14_0\/v1\/user-info,requires-body=1,max-size=0,script-path=https://raw.githubusercontent.com/kityeungfe/Auxiliary/main/Script/longfor/longfor-checkin-simple.js
-
-# 定时签到脚本
-龙湖自动签到 = type=cron,cronexp="*/5 * * * *",wake-system=1,script-path=https://raw.githubusercontent.com/kityeungfe/Auxiliary/main/Script/longfor/longfor-checkin-simple.js
-
-# 使用说明：
-# 1. 安装配置后打开龙湖APP并登录
-# 2. 脚本会自动获取并保存认证信息
-# 3. 每天凌晨1点自动执行签到
-# 4. 签到结果会通过通知推送
-# 5. 兼容 Shadowrocket 环境，内联简化版 ToolKit
