+++ Script/longfor/longfor-checkin.sgmodule
@@ -1,21 +1,25 @@
 #!name=龙湖自动签到
 #!desc=龙湖APP自动签到功能模块，包含获取Token和定时任务
 #!system=ios
+# 使用方法：
+# 1. 将此配置导入 Shadowrocket
+# 2. 开启 MITM 和脚本功能
+# 3. 打开龙湖APP登录, 打开“我的”页面获取token
+# 4. 脚本会自动执行签到
 
 [MITM]
-hostname = %APPEND% gw2c-hw-open.longfor.com
+hostname = gw2c-hw-open.longfor.com
 
 [Script]
-# 获取Token脚本
-龙湖签到Token = type=http-response,pattern=^https:\/\/gw2c-hw-open\.longfor\.com\/supera\/member\/api\/bff\/pages\/v1_14_0\/v1\/user-info,requires-body=1,max-size=0,script-path=https://raw.githubusercontent.com/devkityeung/Auxiliary/main/Script/longfor/longfor-checkin.js,argument=,@require=https://raw.githubusercontent.com/lowking/Scripts/master/util/ToolKit.js
+# 获取Token脚本, 获取成功后可以注释
+龙湖签到Token = type=http-response,pattern=^https:\/\/gw2c-hw-open\.longfor\.com\/supera\/member\/api\/bff\/pages\/v1_14_0\/v1\/user-info,requires-body=1,max-size=0,script-path=https://raw.githubusercontent.com/devkityeung/Auxiliary/main/Script/longfor/longfor-checkin.js
 
 # 定时签到脚本
-龙湖自动签到 = type=cron,cronexp="0 9 * * *",wake-system=1,script-path=https://raw.githubusercontent.com/devkityeung/Auxiliary/main/Script/longfor/longfor-checkin.js,argument=,@require=https://raw.githubusercontent.com/lowking/Scripts/master/util/ToolKit.js
+龙湖自动签到 = type=cron,cronexp="0 9 * * *",wake-system=1,script-path=https://raw.githubusercontent.com/devkityeung/Auxiliary/main/Script/longfor/longfor-checkin.js
 
-# 模块使用说明
-# 1. 安装模块后打开龙湖APP并登录
+# 使用说明：
+# 1. 安装配置后打开龙湖APP并登录
 # 2. 脚本会自动获取并保存认证信息
-# 3. 每天凌晨9点自动执行签到
+# 3. 每天9点自动执行签到
 # 4. 签到结果会通过通知推送
-# 5. 使用外部 ToolKit 库，避免代码冗余
-# 6. 修复了内联代码导致的执行问题
+# 5. 兼容 Shadowrocket 环境，内联简化版 ToolKit
